+++ packages/backend-deployer/src/cdk_deployer.ts
@@ -7,10 +7,11 @@ import {
 } from './cdk_deployer_singleton_factory.js';
 import { CdkErrorMapper } from './cdk_error_mapper.js';
 import { UniqueBackendIdentifier } from '@aws-amplify/plugin-types';
-import { BackendDeploymentType } from '@aws-amplify/platform-core';
+import {
+  BackendDeploymentType,
+  BackendLocator,
+} from '@aws-amplify/platform-core';
 import { BackendDeployerEnvironmentVariables } from './environment_variables.js';
-
-const relativeBackendEntryPoint = 'amplify/backend.ts';
 
 /**
  * Commands that can be invoked
@@ -27,7 +28,10 @@ export class CDKDeployer implements BackendDeployer {
   /**
    * Instantiates instance of CDKDeployer
    */
-  constructor(private readonly cdkErrorMapper: CdkErrorMapper) {}
+  constructor(
+    private readonly cdkErrorMapper: CdkErrorMapper,
+    private readonly backendLocator: BackendLocator
+  ) {}
   /**
    * Invokes cdk deploy command
    */
@@ -93,7 +97,7 @@ export class CDKDeployer implements BackendDeployer {
         'node16',
         '--target',
         'es2022',
-        relativeBackendEntryPoint,
+        this.backendLocator.locate(),
       ]);
     }
   };
@@ -115,7 +119,7 @@ export class CDKDeployer implements BackendDeployer {
       // See https://github.com/aws/aws-cdk/issues/7717 for more details.
       '--ci',
       '--app',
-      `'npx tsx ${relativeBackendEntryPoint}'`,
+      `'npx tsx ${this.backendLocator.locate()}'`,
       '--all',
       '--output',
       '.amplify/artifacts/cdk.out',