+++ packages/integration-tests/src/cdk_out_dir_validator.ts
@@ -17,6 +17,13 @@ const matchHashedJsonFile: Predicate = (actual, expected) => {
   );
 };
 
+const matchTimestamp: Predicate = (actual, expected) =>
+  typeof actual === 'number' && typeof expected === 'number';
+
+const nestedStackDescriptions = new Set([
+  'An auto-generated nested stack for the @function directive.',
+]);
+
 const customMatchers: Map<ObjectPath, Predicate> = new Map([
   [
     [
@@ -69,62 +76,20 @@ const customMatchers: Map<ObjectPath, Predicate> = new Map([
   [
     [
       'Resources',
-      'testGoogleIdSecretFetcherResource',
+      'amplifyDataGraphQLAPIDefaultApiKey1C8ED374',
       'Properties',
-      'secretLastUpdated',
+      'Expires',
     ],
-    (actual) => typeof actual === 'number',
-  ],
-  [
-    [
-      'Resources',
-      'testGoogleSecretSecretFetcherResource',
-      'Properties',
-      'secretLastUpdated',
-    ],
-    (actual) => typeof actual === 'number',
-  ],
-  [
-    [
-      'Resources',
-      'testFacebookIdSecretFetcherResource',
-      'Properties',
-      'secretLastUpdated',
-    ],
-    (actual) => typeof actual === 'number',
-  ],
-  [
-    [
-      'Resources',
-      'testFacebookSecretSecretFetcherResource',
-      'Properties',
-      'secretLastUpdated',
-    ],
-    (actual) => typeof actual === 'number',
-  ],
-  [
-    [
-      'Resources',
-      'testAmazonIdSecretFetcherResource',
-      'Properties',
-      'secretLastUpdated',
-    ],
-    (actual) => typeof actual === 'number',
-  ],
-  [
-    [
-      'Resources',
-      'testAmazonSecretSecretFetcherResource',
-      'Properties',
-      'secretLastUpdated',
-    ],
-    (actual) => typeof actual === 'number',
+    matchTimestamp,
   ],
   [
     ['Description'],
     // the description field of the gql template contains a JSON string that includes "createdOn": "Linux|Mac|Windows"
     // this check just verifies that the string is valid JSON because the createdOn value is different for each platform
-    (actual) => typeof actual === 'string' && JSON.parse(actual),
+    // Deeply nested stacks may not adhere to this, so we check if the stack description is expected, short-circuiting parse.
+    (actual) =>
+      typeof actual === 'string' &&
+      (nestedStackDescriptions.has(actual) || JSON.parse(actual)),
   ],
 ]);