+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/backend-function
 
+## 0.1.4
+
+### Patch Changes
+
+- 68dc91e3: chore: support for JS backend apps
+
+## 0.1.3
+
+### Patch Changes
+
+- 0bd8a3f3: add missing dev deps
+
 ## 0.1.2
 
 ### Patch Changes