+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/backend-cli
 
+## 0.4.0
+
+### Minor Changes
+
+- 6ff17d64: Add 'profile' option to applicable commands
+
+### Patch Changes
+
+- Updated dependencies [68dc91e3]
+  - @aws-amplify/platform-core@0.1.4
+  - @aws-amplify/sandbox@0.2.4
+
+## 0.3.1
+
+### Patch Changes
+
+- 4ad1db95: Gracefully handle Crtl+C error during CLI prompting
+
 ## 0.3.0
 
 ### Minor Changes