+++ packages/integration-tests/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/integration-tests
 
+## 0.2.3
+
+### Patch Changes
+
+- 68dc91e3: chore: support for JS backend apps
+- a126d8df: bump amplify data versions
+
 ## 0.2.2
 
 ### Patch Changes